linux optimization: frtuncate() while mapped
The main point of this change is to avoid msync() when resizing the
file. It turns out remapping is not even needed.

Disclaimer:

> If the size of the mapped file changes after the call to mmap() as a
> result of some other operation on the mapped file, the effect of
> references to portions of the mapped region that correspond to added
> or removed portions of the file is unspecified.

On linux this is a SIGBUS, but by contract of the API the user shouldn't
be writing past the resized area anyway.

An alternative could be to reserve with anonymous+private and mmap()
just the grown region like ResizableMappedMemory. However, mmap is quite
slow, even incrementally.